Refactor app structure into route modules
