+++ config.py
@@ -1,0 +1,3 @@
+class Config:
+    """Application configuration."""
+
