+++ models.py
@@ -1,0 +1,1 @@
+"""Database models will live here when persistence is connected."""
