+++ routes/admin.py
@@ -1,0 +1,1 @@
+"""Admin routes can be added here without crowding the public site app factory."""
